scheduler/requirements.txt
schedule==0.6.0
paho-mqtt==1.3.1